requirements.txt
appdirs==1.4.4
black==23.1.0
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.94.0
flake8==6.0.0
h11==0.14.0
idna==3.4
mccabe==0.7.0
motor==3.1.1
mypy-extensions==1.0.0
pathspec==0.11.0
pycodestyle==2.10.0
pydantic==1.10.6
pyflakes==3.0.1
pymongo==4.3.3
regex==2022.10.31
requests==2.31.0
starlette==0.26.1
toml==0.10.2
typed-ast==1.5.4
typing-extensions==4.5.0
uvicorn==0.21.0
python-dotenv==1.0.0